fix: rename & move user-settings enums to 'types' folder
